Convert snippet preview controller to TypeScript
